requirements.txt
Flask==2.3.0
pandas==2.0.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0
